Unstable: recursion investigation

+++ classes/Options.php
@@ -204,7 +204,9 @@ class Options
             return update_option('_moj_bulk_index_stats', $this->stats_default);
         }
 
-        return unlink($this->importLocation() . 'moj-bulk-index-stats.json');
+        if (file_exists($this->importLocation() . 'moj-bulk-index-stats.json')) {
+            unlink($this->importLocation() . 'moj-bulk-index-stats.json');
+        }
     }
 
     /**
